homepage: Sort language names alphabetically
When the French translation for the homepage was added, the button was
inserted before the button for the Chinese version, to keep the codes
('/fr-fr/', '/zh-cn/') alphabetically sorted in the source. This is
wrong for the list of languages, because 'Chinese' should come before
'French'.

Let's fix this. To avoid future mistakes, let's change the structure
holding the codes and names for the languages, and let's sort it based
on the names. Also use this structure to automatically generate the
list, to avoid writing the names at multiple places.