Add Major/Emerging labels (#153)
This is the first step in a series of changes, split to ease code
review.

Applications already had two sections, which unfortunately
are not adjacent on the page, so this PR just adds the Applications
label to the two sections.  This could have moved the sections around
to put them adjacent but that would have made this PR harder to review,
so leaving it for a separate PR.

The other sections did not have any major/emerging distinction, so this adds
a grouping within each other category.  Currently however the stylesheet
only has two levels of TitleWithAnchor headings where we really need three.
There is currently an inconsistency in terms of style where application
projects use one style and eBPF library projects use a different style.
This PR does not change that yet either, it merely adds labels within
the style that exists for each category.  A future PR should create a
third level and update the styles to be consistent.

Signed-off-by: Dave Thaler <dthaler@microsoft.com>