Reword Projects to Landscape
The use of the word "eBPF Projects" has raised questions whether the
listed projects require to be part of the eBPF foundation somehow. As
per charter, this is not required. The intent is to provide a landscape
of projects.

Reword the page to "eBPF Landscape" to make this clearer.

The URL is kept intact for now to not break any links. A follow-up
change can change non-content mentions of projects and set up an
URL forward.

Signed-off-by: Thomas Graf <thomas@cilium.io>