index: Enlarge the main buttons on the welcome page
The two main buttons (“Whats is eBPF?” and “Project Landscape”) on the
index page had a fixed width of 300px. Since the text on the one on the
right has been changed to “Project Landscape”, the width is not enough
to accommodate for the text length and the padding for the button on my
browser.

This commit raises the width to 350px instead.

Signed-off-by: Quentin Monnet <quentin@isovalent.com>