chore(applications): Turn landscape apps data into JSON
Turn the information about eBPF applications (major and emerging) into
JSON files (formatted with jq, for consistency).

This is so that this data is easier to process, for example to implement
a workflow to automatically update GitHub stars numbers.

Two projects, namely alaz and eunomia-bpf, had no stars count: I added
the numbers in this commit.

The rest of the information should remain the same for all projects; any
other difference introduced by this commit is a bug.

Signed-off-by: Quentin Monnet <qmo@qmon.net>